requirements.txt
gradio==5.31.0
pydantic==2.10.6
opencv-python-headless
einops
einx
plotly
prettytable
decord
easydict
flow_vis
moviepy==1.0.0
safetensors
scikit-learn
hydra-core
omegaconf
albumentations
matplotlib
mediapy
scikit-image
pycolmap==3.11.1
git+https://github.com/facebookresearch/segment-anything.git
git+https://github.com/EasternJournalist/utils3d.git#egg=utils3d
huggingface_hub
pyceres==2.4
kornia==0.8.1
xformers==0.0.28
timm
PyJWT
gdown
rich
decord
ray[default]
jaxtyping
transformers
rerun-sdk